Move copy partner app attachments to separate function and add to task
