+++ EquiTrack/partners/tasks.py
@@ -10,6 +10,7 @@ from celery.utils.log import get_task_logger
 
 from EquiTrack.celery import app
 from partners.models import Agreement, Intervention
+from partners.utils import copy_all_attachments
 from partners.validation.agreements import AgreementValid
 from partners.validation.interventions import InterventionValid
 from users.models import Country, User
@@ -241,3 +242,9 @@ def _notify_interventions_ending_soon(country_name):
             template_data=email_context
         )
         notification.send_notification()
+
+
+@app.task
+def copy_attachments():
+    """Copy all partner app attachments"""
+    copy_all_attachments()
